backend/requirements.txt
fastapi
uvicorn
python-dotenv
azure-identity
azure-storage-blob
imageio-ffmpeg
xmltodict
